requirements.txt
requests
snowflake-connector-python
pandas
python-dotenv
pyarrow==10.0.1
apache-airflow-providers-snowflake